adjust to Fairway package naming conventions, group/artifact id's (#7)
refactor model classes to avoid too many large nested classes
