+++ messaging/src/main/java/org/opentestsystem/rdw/messaging/RdwMessageHeaderAccessor.java
@@ -1,4 +1,4 @@
-package rdw.messaging;
+package org.opentestsystem.rdw.messaging;
 
 import org.springframework.messaging.Message;
 import org.springframework.messaging.support.MessageHeaderAccessor;
