+++ messaging/src/main/java/org/opentestsystem/rdw/messaging/RdwMessageHeaders.java
@@ -1,6 +1,4 @@
-package rdw.messaging;
-
-import org.springframework.messaging.MessageHeaders;
+package org.opentestsystem.rdw.messaging;
 
 /**
  * Pre-defined names and prefixes to be used for setting and/or retrieving RDW message properties
