+++ messaging/src/test/java/org/opentestsystem/rdw/messaging/RdwMessageHeaderAccessorTest.java
@@ -1,4 +1,4 @@
-package rdw.messaging;
+package org.opentestsystem.rdw.messaging;
 
 import org.junit.Test;
 import org.springframework.messaging.Message;
@@ -7,7 +7,7 @@ import org.springframework.util.InvalidMimeTypeException;
 import org.springframework.util.MimeType;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static rdw.messaging.RdwMessageHeaderAccessor.wrap;
+import static org.opentestsystem.rdw.messaging.RdwMessageHeaderAccessor.wrap;
 
 public class RdwMessageHeaderAccessorTest {
