+++ model/src/main/java/org/opentestsystem/rdw/model/Context.java
@@ -1,5 +1,5 @@
 
-package rdw.model;
+package org.opentestsystem.rdw.model;
 
 import javax.xml.bind.annotation.XmlEnum;
 import javax.xml.bind.annotation.XmlType;
