+++ model/src/main/java/org/opentestsystem/rdw/model/Examinee.java
@@ -1,0 +1,444 @@
+package org.opentestsystem.rdw.model;
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlAttribute;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlElements;
+import javax.xml.bind.annotation.XmlSchemaType;
+import javax.xml.bind.annotation.XmlType;
+import javax.xml.datatype.XMLGregorianCalendar;
+import java.math.BigInteger;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * <p>Java class for anonymous complex type.
+ *
+ * <p>The following schema fragment specifies the expected content contained within this class.
+ *
+ * <pre>
+ * &lt;complexType>
+ *   &lt;complexContent>
+ *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *       &lt;choice maxOccurs="unbounded" minOccurs="0">
+ *         &lt;element name="ExamineeAttribute">
+ *           &lt;complexType>
+ *             &lt;complexContent>
+ *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                 &lt;attribute name="context" use="required" type="{}Context" />
+ *                 &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="value" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="contextDate" use="required" type="{http://www.w3.org/2001/XMLSchema}dateTime" />
+ *               &lt;/restriction>
+ *             &lt;/complexContent>
+ *           &lt;/complexType>
+ *         &lt;/element>
+ *         &lt;element name="ExamineeRelationship">
+ *           &lt;complexType>
+ *             &lt;complexContent>
+ *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                 &lt;attribute name="context" use="required" type="{}Context" />
+ *                 &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="entityKey" type="{http://www.w3.org/2001/XMLSchema}unsignedLong" />
+ *                 &lt;attribute name="value" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="contextDate" use="required" type="{http://www.w3.org/2001/XMLSchema}dateTime" />
+ *               &lt;/restriction>
+ *             &lt;/complexContent>
+ *           &lt;/complexType>
+ *         &lt;/element>
+ *       &lt;/choice>
+ *       &lt;attribute name="key" use="required" type="{http://www.w3.org/2001/XMLSchema}long" />
+ *       &lt;attribute name="isDemo" type="{}Bit" />
+ *     &lt;/restriction>
+ *   &lt;/complexContent>
+ * &lt;/complexType>
+ * </pre>
+ *
+ *
+ */
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "", propOrder = {
+    "examineeAttributeOrExamineeRelationship"
+})
+public class Examinee {
+
+    @XmlElements({
+        @XmlElement(name = "ExamineeAttribute", type = ExamineeAttribute.class),
+        @XmlElement(name = "ExamineeRelationship", type = ExamineeRelationship.class)
+    })
+    private List<Object> examineeAttributeOrExamineeRelationship;
+    @XmlAttribute(name = "key", required = true)
+    private long key;
+    @XmlAttribute(name = "isDemo")
+    private Short isDemo;
+
+    /**
+     * Gets the value of the examineeAttributeOrExamineeRelationship property.
+     *
+     * <p>
+     * This accessor method returns a reference to the live list,
+     * not a snapshot. Therefore any modification you make to the
+     * returned list will be present inside the JAXB object.
+     * This is why there is not a <CODE>set</CODE> method for the examineeAttributeOrExamineeRelationship property.
+     *
+     * <p>
+     * For example, to add a new item, do as follows:
+     * <pre>
+     *    getExamineeAttributeOrExamineeRelationship().add(newItem);
+     * </pre>
+     *
+     *
+     * <p>
+     * Objects of the following type(s) are allowed in the list
+     * {@link ExamineeAttribute }
+     * {@link ExamineeRelationship }
+     *
+     *
+     */
+    public List<Object> getExamineeAttributeOrExamineeRelationship() {
+        if (examineeAttributeOrExamineeRelationship == null) {
+            examineeAttributeOrExamineeRelationship = new ArrayList<Object>();
+        }
+        return this.examineeAttributeOrExamineeRelationship;
+    }
+
+    /**
+     * Gets the value of the key property.
+     *
+     */
+    public long getKey() {
+        return key;
+    }
+
+    /**
+     * Sets the value of the key property.
+     *
+     */
+    public void setKey(long value) {
+        this.key = value;
+    }
+
+    /**
+     * Gets the value of the isDemo property.
+     *
+     * @return
+     *     possible object is
+     *     {@link Short }
+     *
+     */
+    public Short getIsDemo() {
+        return isDemo;
+    }
+
+    /**
+     * Sets the value of the isDemo property.
+     *
+     * @param value
+     *     allowed object is
+     *     {@link Short }
+     *
+     */
+    public void setIsDemo(Short value) {
+        this.isDemo = value;
+    }
+
+
+    /**
+     * <p>Java class for anonymous complex type.
+     *
+     * <p>The following schema fragment specifies the expected content contained within this class.
+     *
+     * <pre>
+     * &lt;complexType>
+     *   &lt;complexContent>
+     *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+     *       &lt;attribute name="context" use="required" type="{}Context" />
+     *       &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+     *       &lt;attribute name="value" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+     *       &lt;attribute name="contextDate" use="required" type="{http://www.w3.org/2001/XMLSchema}dateTime" />
+     *     &lt;/restriction>
+     *   &lt;/complexContent>
+     * &lt;/complexType>
+     * </pre>
+     *
+     *
+     */
+    @XmlAccessorType(XmlAccessType.FIELD)
+    @XmlType(name = "")
+    public static class ExamineeAttribute {
+
+        @XmlAttribute(name = "context", required = true)
+        private Context context;
+        @XmlAttribute(name = "name", required = true)
+        @XmlSchemaType(name = "anySimpleType")
+        private String name;
+        @XmlAttribute(name = "value")
+        @XmlSchemaType(name = "anySimpleType")
+        private String value;
+        @XmlAttribute(name = "contextDate", required = true)
+        @XmlSchemaType(name = "dateTime")
+        private XMLGregorianCalendar contextDate;
+
+        /**
+         * Gets the value of the context property.
+         *
+         * @return
+         *     possible object is
+         *     {@link Context }
+         *
+         */
+        public Context getContext() {
+            return context;
+        }
+
+        /**
+         * Sets the value of the context property.
+         *
+         * @param value
+         *     allowed object is
+         *     {@link Context }
+         *
+         */
+        public void setContext(Context value) {
+            this.context = value;
+        }
+
+        /**
+         * Gets the value of the name property.
+         *
+         * @return
+         *     possible object is
+         *     {@link String }
+         *
+         */
+        public String getName() {
+            return name;
+        }
+
+        /**
+         * Sets the value of the name property.
+         *
+         * @param value
+         *     allowed object is
+         *     {@link String }
+         *
+         */
+        public void setName(String value) {
+            this.name = value;
+        }
+
+        /**
+         * Gets the value of the value property.
+         *
+         * @return
+         *     possible object is
+         *     {@link String }
+         *
+         */
+        public String getValue() {
+            return value;
+        }
+
+        /**
+         * Sets the value of the value property.
+         *
+         * @param value
+         *     allowed object is
+         *     {@link String }
+         *
+         */
+        public void setValue(String value) {
+            this.value = value;
+        }
+
+        /**
+         * Gets the value of the contextDate property.
+         *
+         * @return
+         *     possible object is
+         *     {@link XMLGregorianCalendar }
+         *
+         */
+        public XMLGregorianCalendar getContextDate() {
+            return contextDate;
+        }
+
+        /**
+         * Sets the value of the contextDate property.
+         *
+         * @param value
+         *     allowed object is
+         *     {@link XMLGregorianCalendar }
+         *
+         */
+        public void setContextDate(XMLGregorianCalendar value) {
+            this.contextDate = value;
+        }
+
+    }
+
+
+    /**
+     * <p>Java class for anonymous complex type.
+     *
+     * <p>The following schema fragment specifies the expected content contained within this class.
+     *
+     * <pre>
+     * &lt;complexType>
+     *   &lt;complexContent>
+     *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+     *       &lt;attribute name="context" use="required" type="{}Context" />
+     *       &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+     *       &lt;attribute name="entityKey" type="{http://www.w3.org/2001/XMLSchema}unsignedLong" />
+     *       &lt;attribute name="value" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+     *       &lt;attribute name="contextDate" use="required" type="{http://www.w3.org/2001/XMLSchema}dateTime" />
+     *     &lt;/restriction>
+     *   &lt;/complexContent>
+     * &lt;/complexType>
+     * </pre>
+     *
+     *
+     */
+    @XmlAccessorType(XmlAccessType.FIELD)
+    @XmlType(name = "")
+    public static class ExamineeRelationship {
+
+        @XmlAttribute(name = "context", required = true)
+        private Context context;
+        @XmlAttribute(name = "name", required = true)
+        @XmlSchemaType(name = "anySimpleType")
+        private String name;
+        @XmlAttribute(name = "entityKey")
+        @XmlSchemaType(name = "unsignedLong")
+        private BigInteger entityKey;
+        @XmlAttribute(name = "value")
+        @XmlSchemaType(name = "anySimpleType")
+        private String value;
+        @XmlAttribute(name = "contextDate", required = true)
+        @XmlSchemaType(name = "dateTime")
+        private XMLGregorianCalendar contextDate;
+
+        /**
+         * Gets the value of the context property.
+         *
+         * @return
+         *     possible object is
+         *     {@link Context }
+         *
+         */
+        public Context getContext() {
+            return context;
+        }
+
+        /**
+         * Sets the value of the context property.
+         *
+         * @param value
+         *     allowed object is
+         *     {@link Context }
+         *
+         */
+        public void setContext(Context value) {
+            this.context = value;
+        }
+
+        /**
+         * Gets the value of the name property.
+         *
+         * @return
+         *     possible object is
+         *     {@link String }
+         *
+         */
+        public String getName() {
+            return name;
+        }
+
+        /**
+         * Sets the value of the name property.
+         *
+         * @param value
+         *     allowed object is
+         *     {@link String }
+         *
+         */
+        public void setName(String value) {
+            this.name = value;
+        }
+
+        /**
+         * Gets the value of the entityKey property.
+         *
+         * @return
+         *     possible object is
+         *     {@link BigInteger }
+         *
+         */
+        public BigInteger getEntityKey() {
+            return entityKey;
+        }
+
+        /**
+         * Sets the value of the entityKey property.
+         *
+         * @param value
+         *     allowed object is
+         *     {@link BigInteger }
+         *
+         */
+        public void setEntityKey(BigInteger value) {
+            this.entityKey = value;
+        }
+
+        /**
+         * Gets the value of the value property.
+         *
+         * @return
+         *     possible object is
+         *     {@link String }
+         *
+         */
+        public String getValue() {
+            return value;
+        }
+
+        /**
+         * Sets the value of the value property.
+         *
+         * @param value
+         *     allowed object is
+         *     {@link String }
+         *
+         */
+        public void setValue(String value) {
+            this.value = value;
+        }
+
+        /**
+         * Gets the value of the contextDate property.
+         *
+         * @return
+         *     possible object is
+         *     {@link XMLGregorianCalendar }
+         *
+         */
+        public XMLGregorianCalendar getContextDate() {
+            return contextDate;
+        }
+
+        /**
+         * Sets the value of the contextDate property.
+         *
+         * @param value
+         *     allowed object is
+         *     {@link XMLGregorianCalendar }
+         *
+         */
+        public void setContextDate(XMLGregorianCalendar value) {
+            this.contextDate = value;
+        }
+
+    }
+
+}
