+++ model/src/main/java/org/opentestsystem/rdw/model/ObjectFactory.java
@@ -1,0 +1,176 @@
+
+package org.opentestsystem.rdw.model;
+
+import javax.xml.bind.annotation.XmlRegistry;
+
+
+/**
+ * This object contains factory methods for each 
+ * Java content interface and Java element interface 
+ * generated in the rdw.model package. 
+ * <p>An ObjectFactory allows you to programatically 
+ * construct new instances of the Java representation 
+ * for XML content. The Java representation of XML 
+ * content can consist of schema derived interfaces 
+ * and classes representing the binding of schema 
+ * type definitions, element declarations and model 
+ * groups.  Factory methods for each of these are 
+ * provided in this class.
+ * 
+ */
+@XmlRegistry
+public class ObjectFactory {
+
+
+    /**
+     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: rdw.model
+     * 
+     */
+    public ObjectFactory() {
+    }
+
+    /**
+     * Create an instance of {@link TDSReport }
+     * 
+     */
+    public TDSReport createTDSReport() {
+        return new TDSReport();
+    }
+
+    /**
+     * Create an instance of {@link ScoreInfoType }
+     * 
+     */
+    public ScoreInfoType createScoreInfoType() {
+        return new ScoreInfoType();
+    }
+
+    /**
+     * Create an instance of {@link ToolUsage }
+     * 
+     */
+    public ToolUsage createTDSReportToolUsage() {
+        return new ToolUsage();
+    }
+
+    /**
+     * Create an instance of {@link Opportunity }
+     * 
+     */
+    public Opportunity createTDSReportOpportunity() {
+        return new Opportunity();
+    }
+
+    /**
+     * Create an instance of {@link Opportunity.Item }
+     * 
+     */
+    public Opportunity.Item createTDSReportOpportunityItem() {
+        return new Opportunity.Item();
+    }
+
+    /**
+     * Create an instance of {@link Examinee }
+     * 
+     */
+    public Examinee createTDSReportExaminee() {
+        return new Examinee();
+    }
+
+    /**
+     * Create an instance of {@link Test }
+     * 
+     */
+    public Test createTDSReportTest() {
+        return new Test();
+    }
+
+    /**
+     * Create an instance of {@link Comment }
+     * 
+     */
+    public Comment createTDSReportComment() {
+        return new Comment();
+    }
+
+    /**
+     * Create an instance of {@link ScoreInfoType.ScoreRationale }
+     * 
+     */
+    public ScoreInfoType.ScoreRationale createScoreInfoTypeScoreRationale() {
+        return new ScoreInfoType.ScoreRationale();
+    }
+
+    /**
+     * Create an instance of {@link ScoreInfoType.SubScoreList }
+     * 
+     */
+    public ScoreInfoType.SubScoreList createScoreInfoTypeSubScoreList() {
+        return new ScoreInfoType.SubScoreList();
+    }
+
+    /**
+     * Create an instance of {@link ToolUsage.ToolPage }
+     * 
+     */
+    public ToolUsage.ToolPage createTDSReportToolUsageToolPage() {
+        return new ToolUsage.ToolPage();
+    }
+
+    /**
+     * Create an instance of {@link Opportunity.Segment }
+     * 
+     */
+    public Opportunity.Segment createTDSReportOpportunitySegment() {
+        return new Opportunity.Segment();
+    }
+
+    /**
+     * Create an instance of {@link Opportunity.Accommodation }
+     * 
+     */
+    public Opportunity.Accommodation createTDSReportOpportunityAccommodation() {
+        return new Opportunity.Accommodation();
+    }
+
+    /**
+     * Create an instance of {@link Opportunity.Score }
+     * 
+     */
+    public Opportunity.Score createTDSReportOpportunityScore() {
+        return new Opportunity.Score();
+    }
+
+    /**
+     * Create an instance of {@link Opportunity.GenericVariable }
+     * 
+     */
+    public Opportunity.GenericVariable createTDSReportOpportunityGenericVariable() {
+        return new Opportunity.GenericVariable();
+    }
+
+    /**
+     * Create an instance of {@link Opportunity.Item.Response }
+     * 
+     */
+    public Opportunity.Item.Response createTDSReportOpportunityItemResponse() {
+        return new Opportunity.Item.Response();
+    }
+
+    /**
+     * Create an instance of {@link Examinee.ExamineeAttribute }
+     * 
+     */
+    public Examinee.ExamineeAttribute createTDSReportExamineeExamineeAttribute() {
+        return new Examinee.ExamineeAttribute();
+    }
+
+    /**
+     * Create an instance of {@link Examinee.ExamineeRelationship }
+     * 
+     */
+    public Examinee.ExamineeRelationship createTDSReportExamineeExamineeRelationship() {
+        return new Examinee.ExamineeRelationship();
+    }
+
+}
