+++ model/src/main/java/org/opentestsystem/rdw/model/ScoreInfoType.java
@@ -1,5 +1,5 @@
 
-package rdw.model;
+package org.opentestsystem.rdw.model;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -75,22 +75,22 @@ import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 public class ScoreInfoType {
 
     @XmlElement(name = "ScoreRationale")
-    protected ScoreInfoType.ScoreRationale scoreRationale;
+    private ScoreInfoType.ScoreRationale scoreRationale;
     @XmlElement(name = "SubScoreList")
-    protected ScoreInfoType.SubScoreList subScoreList;
+    private ScoreInfoType.SubScoreList subScoreList;
     @XmlAttribute(name = "scorePoint")
-    protected String scorePoint;
+    private String scorePoint;
     @XmlAttribute(name = "maxScore")
-    protected String maxScore;
+    private String maxScore;
     @XmlAttribute(name = "scoreDimension")
     @XmlSchemaType(name = "anySimpleType")
-    protected String scoreDimension;
+    private String scoreDimension;
     @XmlAttribute(name = "scoreStatus")
     @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-    protected String scoreStatus;
+    private String scoreStatus;
     @XmlAttribute(name = "confLevel")
     @XmlSchemaType(name = "anySimpleType")
-    protected String confLevel;
+    private String confLevel;
 
     /**
      * Gets the value of the scoreRationale property.
@@ -287,7 +287,7 @@ public class ScoreInfoType {
     public static class ScoreRationale {
 
         @XmlElement(name = "Message")
-        protected Object message;
+        private Object message;
 
         /**
          * Gets the value of the message property.
@@ -342,7 +342,7 @@ public class ScoreInfoType {
     public static class SubScoreList {
 
         @XmlElement(name = "ScoreInfo")
-        protected List<ScoreInfoType> scoreInfo;
+        private List<ScoreInfoType> scoreInfo;
 
         /**
          * Gets the value of the scoreInfo property.
