+++ model/src/main/java/org/opentestsystem/rdw/model/TDSReport.java
@@ -1,0 +1,462 @@
+
+package org.opentestsystem.rdw.model;
+
+import java.util.ArrayList;
+import java.util.List;
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.XmlType;
+
+
+/**
+ * <p>Java class for anonymous complex type.
+ * 
+ * <p>The following schema fragment specifies the expected content contained within this class.
+ * 
+ * <pre>
+ * &lt;complexType>
+ *   &lt;complexContent>
+ *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *       &lt;sequence>
+ *         &lt;element name="Test">
+ *           &lt;complexType>
+ *             &lt;complexContent>
+ *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                 &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="subject" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="testId" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="bankKey" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                 &lt;attribute name="contract" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="mode" use="required">
+ *                   &lt;simpleType>
+ *                     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}token">
+ *                       &lt;enumeration value="online"/>
+ *                       &lt;enumeration value="paper"/>
+ *                       &lt;enumeration value="scanned"/>
+ *                     &lt;/restriction>
+ *                   &lt;/simpleType>
+ *                 &lt;/attribute>
+ *                 &lt;attribute name="grade" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="handScoreProject" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                 &lt;attribute name="assessmentType" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="academicYear" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                 &lt;attribute name="assessmentVersion" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *               &lt;/restriction>
+ *             &lt;/complexContent>
+ *           &lt;/complexType>
+ *         &lt;/element>
+ *         &lt;element name="Examinee">
+ *           &lt;complexType>
+ *             &lt;complexContent>
+ *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                 &lt;choice maxOccurs="unbounded" minOccurs="0">
+ *                   &lt;element name="ExamineeAttribute">
+ *                     &lt;complexType>
+ *                       &lt;complexContent>
+ *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                           &lt;attribute name="context" use="required" type="{}Context" />
+ *                           &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="value" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="contextDate" use="required" type="{http://www.w3.org/2001/XMLSchema}dateTime" />
+ *                         &lt;/restriction>
+ *                       &lt;/complexContent>
+ *                     &lt;/complexType>
+ *                   &lt;/element>
+ *                   &lt;element name="ExamineeRelationship">
+ *                     &lt;complexType>
+ *                       &lt;complexContent>
+ *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                           &lt;attribute name="context" use="required" type="{}Context" />
+ *                           &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="entityKey" type="{http://www.w3.org/2001/XMLSchema}unsignedLong" />
+ *                           &lt;attribute name="value" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="contextDate" use="required" type="{http://www.w3.org/2001/XMLSchema}dateTime" />
+ *                         &lt;/restriction>
+ *                       &lt;/complexContent>
+ *                     &lt;/complexType>
+ *                   &lt;/element>
+ *                 &lt;/choice>
+ *                 &lt;attribute name="key" use="required" type="{http://www.w3.org/2001/XMLSchema}long" />
+ *                 &lt;attribute name="isDemo" type="{}Bit" />
+ *               &lt;/restriction>
+ *             &lt;/complexContent>
+ *           &lt;/complexType>
+ *         &lt;/element>
+ *         &lt;element name="Opportunity">
+ *           &lt;complexType>
+ *             &lt;complexContent>
+ *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                 &lt;sequence>
+ *                   &lt;element name="Segment" maxOccurs="unbounded" minOccurs="0">
+ *                     &lt;complexType>
+ *                       &lt;complexContent>
+ *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                           &lt;attribute name="id" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="position" use="required">
+ *                             &lt;simpleType>
+ *                               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}unsignedByte">
+ *                                 &lt;minInclusive value="1"/>
+ *                               &lt;/restriction>
+ *                             &lt;/simpleType>
+ *                           &lt;/attribute>
+ *                           &lt;attribute name="formId" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="formKey" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="algorithm" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="algorithmVersion" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                         &lt;/restriction>
+ *                       &lt;/complexContent>
+ *                     &lt;/complexType>
+ *                   &lt;/element>
+ *                   &lt;element name="Accommodation" maxOccurs="unbounded" minOccurs="0">
+ *                     &lt;complexType>
+ *                       &lt;complexContent>
+ *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                           &lt;attribute name="type" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="value" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="code" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="segment" use="required">
+ *                             &lt;simpleType>
+ *                               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}unsignedInt">
+ *                                 &lt;minInclusive value="0"/>
+ *                               &lt;/restriction>
+ *                             &lt;/simpleType>
+ *                           &lt;/attribute>
+ *                         &lt;/restriction>
+ *                       &lt;/complexContent>
+ *                     &lt;/complexType>
+ *                   &lt;/element>
+ *                   &lt;element name="Score" maxOccurs="unbounded" minOccurs="0">
+ *                     &lt;complexType>
+ *                       &lt;complexContent>
+ *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                           &lt;attribute name="measureOf" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="measureLabel" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="value" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="standardError" type="{}NullableFloat" />
+ *                         &lt;/restriction>
+ *                       &lt;/complexContent>
+ *                     &lt;/complexType>
+ *                   &lt;/element>
+ *                   &lt;element name="GenericVariable" maxOccurs="unbounded" minOccurs="0">
+ *                     &lt;complexType>
+ *                       &lt;complexContent>
+ *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                           &lt;attribute name="context" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="name" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="value" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                         &lt;/restriction>
+ *                       &lt;/complexContent>
+ *                     &lt;/complexType>
+ *                   &lt;/element>
+ *                   &lt;element name="Item" maxOccurs="unbounded" minOccurs="0">
+ *                     &lt;complexType>
+ *                       &lt;complexContent>
+ *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                           &lt;sequence>
+ *                             &lt;element name="Response" minOccurs="0">
+ *                               &lt;complexType>
+ *                                 &lt;complexContent>
+ *                                   &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                                     &lt;attribute name="date" type="{http://www.w3.org/2001/XMLSchema}dateTime" />
+ *                                     &lt;attribute name="type">
+ *                                       &lt;simpleType>
+ *                                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}token">
+ *                                           &lt;enumeration value="value"/>
+ *                                           &lt;enumeration value="reference"/>
+ *                                           &lt;enumeration value=""/>
+ *                                         &lt;/restriction>
+ *                                       &lt;/simpleType>
+ *                                     &lt;/attribute>
+ *                                   &lt;/restriction>
+ *                                 &lt;/complexContent>
+ *                               &lt;/complexType>
+ *                             &lt;/element>
+ *                             &lt;element name="ScoreInfo" type="{}ScoreInfoType" minOccurs="0"/>
+ *                           &lt;/sequence>
+ *                           &lt;attribute name="responseDuration" type="{}UFloat" />
+ *                           &lt;attribute name="position" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                           &lt;attribute name="segmentId" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="bankKey" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                           &lt;attribute name="key" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                           &lt;attribute name="operational" use="required" type="{}Bit" />
+ *                           &lt;attribute name="isSelected" use="required" type="{}Bit" />
+ *                           &lt;attribute name="format" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="score" use="required" type="{}UFloatAllowNegativeOne" />
+ *                           &lt;attribute name="scoreStatus">
+ *                             &lt;simpleType>
+ *                               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}token">
+ *                                 &lt;enumeration value="NOTSCORED"/>
+ *                                 &lt;enumeration value="SCORED"/>
+ *                                 &lt;enumeration value="WAITINGFORMACHINESCORE"/>
+ *                                 &lt;enumeration value="SCORINGERROR"/>
+ *                                 &lt;enumeration value="APPEALED"/>
+ *                               &lt;/restriction>
+ *                             &lt;/simpleType>
+ *                           &lt;/attribute>
+ *                           &lt;attribute name="adminDate" use="required" type="{http://www.w3.org/2001/XMLSchema}dateTime" />
+ *                           &lt;attribute name="numberVisits" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                           &lt;attribute name="mimeType" use="required">
+ *                             &lt;simpleType>
+ *                               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}token">
+ *                                 &lt;enumeration value="text/plain"/>
+ *                                 &lt;enumeration value="text/xml"/>
+ *                                 &lt;enumeration value="text/html"/>
+ *                                 &lt;enumeration value="audio/ogg"/>
+ *                               &lt;/restriction>
+ *                             &lt;/simpleType>
+ *                           &lt;/attribute>
+ *                           &lt;attribute name="clientId" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="strand" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="contentLevel" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="pageNumber" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                           &lt;attribute name="pageVisits" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                           &lt;attribute name="pageTime" use="required" type="{http://www.w3.org/2001/XMLSchema}int" />
+ *                           &lt;attribute name="dropped" use="required" type="{}Bit" />
+ *                         &lt;/restriction>
+ *                       &lt;/complexContent>
+ *                     &lt;/complexType>
+ *                   &lt;/element>
+ *                 &lt;/sequence>
+ *                 &lt;attribute name="server" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="database" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="clientName" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="key" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="oppId" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="startDate" type="{}NullableDateTime" />
+ *                 &lt;attribute name="status" use="required">
+ *                   &lt;simpleType>
+ *                     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}token">
+ *                       &lt;enumeration value="appeal"/>
+ *                       &lt;enumeration value="completed"/>
+ *                       &lt;enumeration value="expired"/>
+ *                       &lt;enumeration value="handscoring"/>
+ *                       &lt;enumeration value="invalidated"/>
+ *                       &lt;enumeration value="paused"/>
+ *                       &lt;enumeration value="reported"/>
+ *                       &lt;enumeration value="reset"/>
+ *                       &lt;enumeration value="scored"/>
+ *                       &lt;enumeration value="submitted"/>
+ *                       &lt;enumeration value="pending"/>
+ *                     &lt;/restriction>
+ *                   &lt;/simpleType>
+ *                 &lt;/attribute>
+ *                 &lt;attribute name="opportunity" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                 &lt;attribute name="statusDate" use="required" type="{http://www.w3.org/2001/XMLSchema}dateTime" />
+ *                 &lt;attribute name="dateCompleted" type="{}NullableDateTime" />
+ *                 &lt;attribute name="pauseCount" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                 &lt;attribute name="itemCount" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                 &lt;attribute name="ftCount" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                 &lt;attribute name="abnormalStarts" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                 &lt;attribute name="gracePeriodRestarts" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                 &lt;attribute name="taId" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="taName" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="sessionId" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="windowId" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="windowOpportunity" type="{}NullableUInt" />
+ *                 &lt;attribute name="completeStatus" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="administrationCondition" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="dateForceCompleted" type="{}NullableDateTime" />
+ *                 &lt;attribute name="qaLevel" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="assessmentParticipantSessionPlatformUserAgent" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="effectiveDate" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="reportingVersion" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *               &lt;/restriction>
+ *             &lt;/complexContent>
+ *           &lt;/complexType>
+ *         &lt;/element>
+ *         &lt;element name="Comment" maxOccurs="unbounded" minOccurs="0">
+ *           &lt;complexType>
+ *             &lt;complexContent>
+ *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                 &lt;attribute name="context" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="itemPosition" type="{}NullableUInt" />
+ *                 &lt;attribute name="date" use="required" type="{http://www.w3.org/2001/XMLSchema}dateTime" />
+ *               &lt;/restriction>
+ *             &lt;/complexContent>
+ *           &lt;/complexType>
+ *         &lt;/element>
+ *         &lt;element name="ToolUsage" maxOccurs="unbounded" minOccurs="0">
+ *           &lt;complexType>
+ *             &lt;complexContent>
+ *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                 &lt;sequence>
+ *                   &lt;element name="ToolPage" maxOccurs="unbounded">
+ *                     &lt;complexType>
+ *                       &lt;complexContent>
+ *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                           &lt;attribute name="page" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                           &lt;attribute name="groupId" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                           &lt;attribute name="count" use="required" type="{http://www.w3.org/2001/XMLSchema}unsignedInt" />
+ *                         &lt;/restriction>
+ *                       &lt;/complexContent>
+ *                     &lt;/complexType>
+ *                   &lt;/element>
+ *                 &lt;/sequence>
+ *                 &lt;attribute name="type" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *                 &lt;attribute name="code" use="required" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
+ *               &lt;/restriction>
+ *             &lt;/complexContent>
+ *           &lt;/complexType>
+ *         &lt;/element>
+ *       &lt;/sequence>
+ *     &lt;/restriction>
+ *   &lt;/complexContent>
+ * &lt;/complexType>
+ * </pre>
+ * 
+ * 
+ */
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "", propOrder = {
+    "test",
+    "examinee",
+    "opportunity",
+    "comment",
+    "toolUsage"
+})
+@XmlRootElement(name = "TDSReport")
+public class TDSReport {
+
+    @XmlElement(name = "Test", required = true)
+    private Test test;
+    @XmlElement(name = "Examinee", required = true)
+    private Examinee examinee;
+    @XmlElement(name = "Opportunity", required = true)
+    private Opportunity opportunity;
+    @XmlElement(name = "Comment")
+    private List<Comment> comment;
+    @XmlElement(name = "ToolUsage")
+    private List<ToolUsage> toolUsage;
+
+    /**
+     * Gets the value of the test property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link Test }
+     *     
+     */
+    public Test getTest() {
+        return test;
+    }
+
+    /**
+     * Sets the value of the test property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link Test }
+     *     
+     */
+    public void setTest(Test value) {
+        this.test = value;
+    }
+
+    /**
+     * Gets the value of the examinee property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link Examinee }
+     *     
+     */
+    public Examinee getExaminee() {
+        return examinee;
+    }
+
+    /**
+     * Sets the value of the examinee property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link Examinee }
+     *     
+     */
+    public void setExaminee(Examinee value) {
+        this.examinee = value;
+    }
+
+    /**
+     * Gets the value of the opportunity property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link Opportunity }
+     *     
+     */
+    public Opportunity getOpportunity() {
+        return opportunity;
+    }
+
+    /**
+     * Sets the value of the opportunity property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link Opportunity }
+     *     
+     */
+    public void setOpportunity(Opportunity value) {
+        this.opportunity = value;
+    }
+
+    /**
+     * Gets the value of the comment property.
+     * 
+     * <p>
+     * This accessor method returns a reference to the live list,
+     * not a snapshot. Therefore any modification you make to the
+     * returned list will be present inside the JAXB object.
+     * This is why there is not a <CODE>set</CODE> method for the comment property.
+     * 
+     * <p>
+     * For example, to add a new item, do as follows:
+     * <pre>
+     *    getComment().add(newItem);
+     * </pre>
+     * 
+     * 
+     * <p>
+     * Objects of the following type(s) are allowed in the list
+     * {@link Comment }
+     * 
+     * 
+     */
+    public List<Comment> getComment() {
+        if (comment == null) {
+            comment = new ArrayList<Comment>();
+        }
+        return this.comment;
+    }
+
+    /**
+     * Gets the value of the toolUsage property.
+     * 
+     * <p>
+     * This accessor method returns a reference to the live list,
+     * not a snapshot. Therefore any modification you make to the
+     * returned list will be present inside the JAXB object.
+     * This is why there is not a <CODE>set</CODE> method for the toolUsage property.
+     * 
+     * <p>
+     * For example, to add a new item, do as follows:
+     * <pre>
+     *    getToolUsage().add(newItem);
+     * </pre>
+     * 
+     * 
+     * <p>
+     * Objects of the following type(s) are allowed in the list
+     * {@link ToolUsage }
+     * 
+     * 
+     */
+    public List<ToolUsage> getToolUsage() {
+        if (toolUsage == null) {
+            toolUsage = new ArrayList<ToolUsage>();
+        }
+        return this.toolUsage;
+    }
+}
