+++ model/src/main/java/org/opentestsystem/rdw/model/XmlUtils.java
@@ -1,4 +1,4 @@
-package rdw.model;
+package org.opentestsystem.rdw.model;
 
 import org.xml.sax.SAXException;
 
@@ -45,7 +45,7 @@ public class XmlUtils {
     private static JAXBContext getContext() throws JAXBException {
         // although not thread-safe, the worst that happens is double instantiation
         if (context == null) {
-            context = JAXBContext.newInstance("rdw.model");
+            context = JAXBContext.newInstance("org.opentestsystem.rdw.model");
         }
         return context;
     }
