--- model/src/main/java/rdw/model/ObjectFactory.java
@@ -1,176 +1,0 @@
-
-package rdw.model;
-
-import javax.xml.bind.annotation.XmlRegistry;
-
-
-/**
- * This object contains factory methods for each 
- * Java content interface and Java element interface 
- * generated in the rdw.model package. 
- * <p>An ObjectFactory allows you to programatically 
- * construct new instances of the Java representation 
- * for XML content. The Java representation of XML 
- * content can consist of schema derived interfaces 
- * and classes representing the binding of schema 
- * type definitions, element declarations and model 
- * groups.  Factory methods for each of these are 
- * provided in this class.
- * 
- */
-@XmlRegistry
-public class ObjectFactory {
-
-
-    /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: rdw.model
-     * 
-     */
-    public ObjectFactory() {
-    }
-
-    /**
-     * Create an instance of {@link TDSReport }
-     * 
-     */
-    public TDSReport createTDSReport() {
-        return new TDSReport();
-    }
-
-    /**
-     * Create an instance of {@link ScoreInfoType }
-     * 
-     */
-    public ScoreInfoType createScoreInfoType() {
-        return new ScoreInfoType();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.ToolUsage }
-     * 
-     */
-    public TDSReport.ToolUsage createTDSReportToolUsage() {
-        return new TDSReport.ToolUsage();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Opportunity }
-     * 
-     */
-    public TDSReport.Opportunity createTDSReportOpportunity() {
-        return new TDSReport.Opportunity();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Opportunity.Item }
-     * 
-     */
-    public TDSReport.Opportunity.Item createTDSReportOpportunityItem() {
-        return new TDSReport.Opportunity.Item();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Examinee }
-     * 
-     */
-    public TDSReport.Examinee createTDSReportExaminee() {
-        return new TDSReport.Examinee();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Test }
-     * 
-     */
-    public TDSReport.Test createTDSReportTest() {
-        return new TDSReport.Test();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Comment }
-     * 
-     */
-    public TDSReport.Comment createTDSReportComment() {
-        return new TDSReport.Comment();
-    }
-
-    /**
-     * Create an instance of {@link ScoreInfoType.ScoreRationale }
-     * 
-     */
-    public ScoreInfoType.ScoreRationale createScoreInfoTypeScoreRationale() {
-        return new ScoreInfoType.ScoreRationale();
-    }
-
-    /**
-     * Create an instance of {@link ScoreInfoType.SubScoreList }
-     * 
-     */
-    public ScoreInfoType.SubScoreList createScoreInfoTypeSubScoreList() {
-        return new ScoreInfoType.SubScoreList();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.ToolUsage.ToolPage }
-     * 
-     */
-    public TDSReport.ToolUsage.ToolPage createTDSReportToolUsageToolPage() {
-        return new TDSReport.ToolUsage.ToolPage();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Opportunity.Segment }
-     * 
-     */
-    public TDSReport.Opportunity.Segment createTDSReportOpportunitySegment() {
-        return new TDSReport.Opportunity.Segment();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Opportunity.Accommodation }
-     * 
-     */
-    public TDSReport.Opportunity.Accommodation createTDSReportOpportunityAccommodation() {
-        return new TDSReport.Opportunity.Accommodation();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Opportunity.Score }
-     * 
-     */
-    public TDSReport.Opportunity.Score createTDSReportOpportunityScore() {
-        return new TDSReport.Opportunity.Score();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Opportunity.GenericVariable }
-     * 
-     */
-    public TDSReport.Opportunity.GenericVariable createTDSReportOpportunityGenericVariable() {
-        return new TDSReport.Opportunity.GenericVariable();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Opportunity.Item.Response }
-     * 
-     */
-    public TDSReport.Opportunity.Item.Response createTDSReportOpportunityItemResponse() {
-        return new TDSReport.Opportunity.Item.Response();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Examinee.ExamineeAttribute }
-     * 
-     */
-    public TDSReport.Examinee.ExamineeAttribute createTDSReportExamineeExamineeAttribute() {
-        return new TDSReport.Examinee.ExamineeAttribute();
-    }
-
-    /**
-     * Create an instance of {@link TDSReport.Examinee.ExamineeRelationship }
-     * 
-     */
-    public TDSReport.Examinee.ExamineeRelationship createTDSReportExamineeExamineeRelationship() {
-        return new TDSReport.Examinee.ExamineeRelationship();
-    }
-
-}
