+++ model/src/test/java/org/opentestsystem/rdw/model/XmlUtilsTest.java
@@ -1,4 +1,4 @@
-package rdw.model;
+package org.opentestsystem.rdw.model;
 
 import org.junit.Test;
