+++ utils/build.gradle
@@ -1,6 +1,3 @@
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
-
 dependencies {
     compile ('org.springframework:spring-context:+')
 }
