+++ utils/src/main/java/org/opentestsystem/rdw/utils/TenancyChain.java
@@ -1,4 +1,4 @@
-package rdw.utils;
+package org.opentestsystem.rdw.utils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -45,6 +45,7 @@ public class TenancyChain {
      *
      * @param value chain to parse
      * @return newly created tenancy chain
+     * @see TenancyChain#toString()
      */
     public static TenancyChain fromString(final String value) {
         final ArrayList<Grant> grants = new ArrayList<>();
@@ -61,10 +62,17 @@ public class TenancyChain {
         return new TenancyChain(grants);
     }
 
+    /**
+     * @return tenancy chain as a string
+     * @see TenancyChain#fromString(String)
+     */
     public String toString() {
         return String.join(",", grants.stream().map(Grant::toString).collect(Collectors.toList()));
     }
 
+    /**
+     * @return true if this chain has no role grants
+     */
     public boolean isEmpty() {
         return grants.isEmpty();
     }
@@ -97,7 +105,11 @@ public class TenancyChain {
     }
 
 
-    public static class Grant {
+    /**
+     * A helper class that wraps the field tokens providing named accessors.
+     * It also does basic validation between grant level and field data.
+     */
+    static class Grant {
         private static final int idxRoleId = 1;
         private static final int idxRoleName = 2;
         private static final int idxLevel = 3;
