+++ utils/src/main/java/org/opentestsystem/rdw/utils/YamlPropertiesConfigurator.java
@@ -1,4 +1,4 @@
-package rdw.utils;
+package org.opentestsystem.rdw.utils;
 
 import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
 import org.springframework.context.ApplicationContext;
@@ -13,7 +13,7 @@ import java.util.List;
 import java.util.Properties;
 
 /**
- * Created on 2/22/17.
+ * A {@link Configuration} class for loading SQL YML files as property source.
  */
 @Configuration
 public class YamlPropertiesConfigurator {
