+++ utils/src/test/java/org/opentestsystem/rdw/utils/TenancyChainTest.java
@@ -1,4 +1,4 @@
-package rdw.utils;
+package org.opentestsystem.rdw.utils;
 
 import org.junit.Test;
